requirements.txt
Django==3.0.3
django-cors-headers>=3.0.1
djangorestframework>=3.9.2
Pillow>=6.2.0
psycopg2-binary>=2.8.4,<2.9
djangorestframework-simplejwt>=4.3.0
requests>=2.22

